requirements.txt
py-cord==2.6.1
python-dotenv==1.0.1
pytz==2024.2